+++ src/containerapp-preview/azext_containerapp_preview/tests/latest/test_containerapp_preview_environment.py
@@ -1,0 +1,87 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+import os
+import unittest  # pylint: disable=unused-import
+
+from azure.cli.testsdk import (ScenarioTest, ResourceGroupPreparer)
+
+
+TEST_DIR = os.path.abspath(os.path.join(os.path.abspath(__file__), '..'))
+
+
+class ContainerappComposePreviewEnvironmentSettingsScenarioTest(ScenarioTest):
+    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
+    def test_containerapp_compose_create_with_environment(self, resource_group):
+        compose_text = """
+services:
+  foo:
+    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
+    environment:
+      - RACK_ENV=development
+      - SHOW=true
+      - BAZ="snafu"
+"""
+        compose_file_name = f"{self._testMethodName}_compose.yml"
+        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
+        _ = docker_compose_file.write(compose_text)
+        docker_compose_file.close()
+
+        self.kwargs.update({
+            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
+            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
+            'compose': compose_file_name,
+        })
+
+        command_string = 'containerapp compose create'
+        command_string += ' --compose-file-path {compose}'
+        command_string += ' --resource-group {rg}'
+        command_string += ' --environment {environment}'
+        command_string += ' --logs-workspace {workspace}'
+        self.cmd(command_string, checks=[
+            self.check('[?name==`foo`].properties.template.containers[0].env[0].name', ["RACK_ENV"]),
+            self.check('[?name==`foo`].properties.template.containers[0].env[0].value', ["development"]),
+            self.check('[?name==`foo`].properties.template.containers[0].env[1].name', ["SHOW"]),
+            self.check('[?name==`foo`].properties.template.containers[0].env[1].value', ["true"]),
+            self.check('[?name==`foo`].properties.template.containers[0].env[2].name', ["BAZ"]),
+            self.check('[?name==`foo`].properties.template.containers[0].env[2].value', ['"snafu"'])
+        ])
+
+        if os.path.exists(compose_file_name):
+            os.remove(compose_file_name)
+
+
+class ContainerappComposePreviewEnvironmentSettingsExpectedExceptionScenarioTest(ScenarioTest):
+    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
+    def test_containerapp_compose_create_with_environment_prompt(self, resource_group):
+        compose_text = """
+services:
+  foo:
+    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
+    environment:
+      - LOREM=
+"""
+        compose_file_name = f"{self._testMethodName}_compose.yml"
+        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
+        _ = docker_compose_file.write(compose_text)
+        docker_compose_file.close()
+
+        self.kwargs.update({
+            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
+            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
+            'compose': compose_file_name,
+        })
+
+        command_string = 'containerapp compose create'
+        command_string += ' --compose-file-path {compose}'
+        command_string += ' --resource-group {rg}'
+        command_string += ' --environment {environment}'
+        command_string += ' --logs-workspace {workspace}'
+
+        # This test fails because prompts are not supported in NoTTY environments
+        self.cmd(command_string, expect_failure=True)
+
+        if os.path.exists(compose_file_name):
+            os.remove(compose_file_name)
