--- src/containerapp-preview/azext_containerapp_preview/tests/latest/test_containerapp_preview_scenario.py
@@ -1,821 +1,0 @@
-# --------------------------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License. See License.txt in the project root for license information.
-# --------------------------------------------------------------------------------------------
-
-import os
-import unittest  # pylint: disable=unused-import
-
-from azure.cli.testsdk import (ScenarioTest, ResourceGroupPreparer)
-
-
-TEST_DIR = os.path.abspath(os.path.join(os.path.abspath(__file__), '..'))
-
-
-class ContainerappComposePreviewScenarioTest(ScenarioTest):
-
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_no_existing_resources(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: smurawski/printenv:latest
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-        self.cmd(command_string, checks=[
-            self.check('[].name', ['foo']),
-            self.check('[] | length(@)', 1),
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-
-class ContainerappComposePreviewIngressScenarioTest(ScenarioTest):
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_external_ingress(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    ports: 8080:80
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-        self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.configuration.ingress.targetPort', [80]),
-            self.check('[?name==`foo`].properties.configuration.ingress.external', [True]),
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-
-class ContainerappComposePreviewIngressInternalScenarioTest(ScenarioTest):
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_internal_ingress(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    expose:
-      - "3000"
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-        self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.configuration.ingress.targetPort', [3000]),
-            self.check('[?name==`foo`].properties.configuration.ingress.external', [False]),
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-
-class ContainerappComposePreviewIngressBothScenarioTest(ScenarioTest):
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_both_ingress(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    ports: 4000:3000
-    expose:
-      - "5000"
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-        self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.configuration.ingress.targetPort', [3000]),
-            self.check('[?name==`foo`].properties.configuration.ingress.external', [True]),
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-
-class ContainerappComposePreviewIngressPromptScenarioTest(ScenarioTest):
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_prompt_ingress(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    ports:
-      - 4000:3000
-      - 443:443
-      - 80:80
-    expose:
-      - "5000"
-      - "3000"
-      - "443"
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-
-        # This test fails because prompts are not supported in NoTTY environments
-        self.cmd(command_string, expect_failure=True)
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-
-class ContainerappComposePreviewCommandScenarioTest(ScenarioTest):
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_with_command_string(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    command: echo "hello world"
-    expose:
-      - "5000"
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-        self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.template.containers[0].command[0]', "['echo \"hello world\"']"),
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_with_command_list(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    command: ["echo", "hello world"]
-    expose:
-      - "5000"
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-        self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.template.containers[0].command[0]', "['echo \"hello world\"']"),
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_with_entrypoint_and_command_list(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    command: ["echo", "hello world"]
-    entrypoint: /code/entrypoint.sh
-    expose:
-      - "5000"
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-        self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.template.containers[0].command[0]', "['/code/entrypoint.sh']"),
-            self.check('[?name==`foo`].properties.template.containers[0].args[0]', "['echo \"hello world\"']"),
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-
-class ContainerappComposePreviewResourceSettingsScenarioTest(ScenarioTest):
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_cpus(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    cpus: 1.25
-    expose:
-      - "3000"
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-        self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.template.containers[0].resources.cpu', [1.25]),
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_deploy_cpu(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    deploy:
-      resources:
-        reservations:
-          cpus: 1.25
-    expose:
-      - "3000"
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-        self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.template.containers[0].resources.cpu', [1.25]),
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_cpus_and_deploy_cpu(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    cpus: 0.75
-    deploy:
-      resources:
-        reservations:
-          cpus: 1.25
-    expose:
-      - "3000"
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-        self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.template.containers[0].resources.cpu', [1.25]),
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-
-class ContainerappComposePreviewEnvironmentSettingsScenarioTest(ScenarioTest):
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_environment(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    environment:
-      - RACK_ENV=development
-      - SHOW=true
-      - BAZ="snafu"
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-        self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.template.containers[0].env[0].name', ["RACK_ENV"]),
-            self.check('[?name==`foo`].properties.template.containers[0].env[0].value', ["development"]),
-            self.check('[?name==`foo`].properties.template.containers[0].env[1].name', ["SHOW"]),
-            self.check('[?name==`foo`].properties.template.containers[0].env[1].value', ["true"]),
-            self.check('[?name==`foo`].properties.template.containers[0].env[2].name', ["BAZ"]),
-            self.check('[?name==`foo`].properties.template.containers[0].env[2].value', ['"snafu"'])
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-
-class ContainerappComposePreviewEnvironmentSettingsExpectedExceptionScenarioTest(ScenarioTest):
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_environment_prompt(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    environment:
-      - LOREM=
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-
-        # This test fails because prompts are not supported in NoTTY environments
-        self.cmd(command_string, expect_failure=True)
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-
-class ContainerappComposePreviewTransportOverridesScenarioTest(ScenarioTest):
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_transport_arg(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    ports: 8080:80
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-            'transport': "foo=http2 bar=auto",
-            'second_transport': "baz=http",
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-        command_string += ' --transport {transport}'
-        command_string += ' --transport {second_transport}'
-        self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.configuration.ingress.transport', ["Http2"]),
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-
-class ContainerappComposePreviewRegistryAllArgsScenarioTest(ScenarioTest):
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_all_registry_args(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    ports: 8080:80
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-            'registry_server': "foobar.azurecr.io",
-            'registry_user': "foobar",
-            'registry_pass': "snafu",
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-        command_string += ' --registry-server {registry_server}'
-        command_string += ' --registry-username {registry_user}'
-        command_string += ' --registry-password {registry_pass}'
-
-        self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.configuration.registries[0].server', ["foobar.azurecr.io"]),
-            self.check('[?name==`foo`].properties.configuration.registries[0].username', ["foobar"]),
-            self.check('[?name==`foo`].properties.configuration.registries[0].passwordSecretRef', ["foobarazurecrio-foobar"]),  # pylint: disable=C0301
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-
-class ContainerappComposePreviewRegistryServerArgOnlyScenarioTest(ScenarioTest):
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_registry_server_arg_only(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    ports: 8080:80
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-            'registry_server': "foobar.azurecr.io",
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-        command_string += ' --registry-server {registry_server}'
-
-        # This test fails because prompts are not supported in NoTTY environments
-        self.cmd(command_string, expect_failure=True)
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-
-class ContainerappComposePreviewSecretsScenarioTest(ScenarioTest):
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_secrets(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    secrets:
-      - source: my_secret
-        target: redis_secret
-        uid: '103'
-        gid: '103'
-        mode: 0440
-secrets:
-  my_secret:
-    file: ./my_secret.txt
-  my_other_secret:
-    external: true
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        secrets_file_name = "./my_secret.txt"
-        docker_secrets_file = open(secrets_file_name, "w", encoding='utf-8')
-        _ = docker_secrets_file.write("Lorem Ipsum\n")
-        docker_secrets_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-
-        self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.configuration.secrets[0].name', ["redis-secret"]),
-            self.check('[?name==`foo`].properties.template.containers[0].env[0].name', ["redis-secret"]),
-            self.check('[?name==`foo`].properties.template.containers[0].env[0].secretRef', ["redis-secret"])  # pylint: disable=C0301
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-        if os.path.exists(secrets_file_name):
-            os.remove(secrets_file_name)
-
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_secrets_and_existing_environment(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    environment:
-      database__client: mysql
-      database__connection__host: db
-      database__connection__user: root
-      database__connection__password: example
-      database__connection__database: snafu
-    secrets:
-      - source: my_secret
-        target: redis_secret
-        uid: '103'
-        gid: '103'
-        mode: 0440
-secrets:
-  my_secret:
-    file: ./snafu.txt
-  my_other_secret:
-    external: true
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        secrets_file_name = "./snafu.txt"
-        docker_secrets_file = open(secrets_file_name, "w", encoding='utf-8')
-        _ = docker_secrets_file.write("Lorem Ipsum\n")
-        docker_secrets_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-
-        self.cmd(command_string, checks=[
-            self.check('length([?name==`foo`].properties.template.containers[0].env[].name)', 6),
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-        if os.path.exists(secrets_file_name):
-            os.remove(secrets_file_name)
-
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_secrets_and_existing_environment_conflict(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    environment:
-      database--client: mysql
-    secrets:
-      -  database__client
-secrets:
-  database__client:
-    file: ./database__client.txt
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        secrets_file_name = "./database__client.txt"
-        docker_secrets_file = open(secrets_file_name, "w", encoding='utf-8')
-        _ = docker_secrets_file.write("Lorem Ipsum\n")
-        docker_secrets_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-
-        # This test fails with duplicate environment variable names
-        self.cmd(command_string, expect_failure=True)
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-        if os.path.exists(secrets_file_name):
-            os.remove(secrets_file_name)
-
-
-class ContainerappComposePreviewReplicasScenarioTest(ScenarioTest):
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_replicas_global_scale(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    ports: 8080:80
-    scale: 4
-    deploy:
-      mode: global
-      replicas: 6
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-
-        self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.template.scale.minReplicas', [1]),
-            self.check('[?name==`foo`].properties.template.scale.maxReplicas', [1]),
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
-
-    @ResourceGroupPreparer(name_prefix='cli_test_containerapp_preview', location='eastus')
-    def test_containerapp_compose_create_with_replicas_replicated_mode(self, resource_group):
-        compose_text = """
-services:
-  foo:
-    image: mcr.microsoft.com/azuredocs/aks-helloworld:v1
-    ports: 8080:80
-    deploy:
-      mode: replicated
-      replicas: 6
-"""
-        compose_file_name = f"{self._testMethodName}_compose.yml"
-        docker_compose_file = open(compose_file_name, "w", encoding='utf-8')
-        _ = docker_compose_file.write(compose_text)
-        docker_compose_file.close()
-
-        self.kwargs.update({
-            'environment': self.create_random_name(prefix='containerapp-preview', length=24),
-            'workspace': self.create_random_name(prefix='containerapp-preview', length=24),
-            'compose': compose_file_name,
-        })
-
-        command_string = 'containerapp compose create'
-        command_string += ' --compose-file-path {compose}'
-        command_string += ' --resource-group {rg}'
-        command_string += ' --environment {environment}'
-        command_string += ' --logs-workspace {workspace}'
-
-        self.cmd(command_string, checks=[
-            self.check('[?name==`foo`].properties.template.scale.minReplicas', [6]),
-            self.check('[?name==`foo`].properties.template.scale.maxReplicas', [6]),
-        ])
-
-        if os.path.exists(compose_file_name):
-            os.remove(compose_file_name)
